Added the MarkerTarget object files.
I also changed the ObjectType enum a bit since there will be four markers, and I don't want
to add a field in ObjectData to tell which marker was detected.

I made the destructor for VisibleObject virtual. This is to ensure proper destruction when
deleting a polymorphic VisibleObject pointer. If the destructor is not virtual, only the
destructor for VisibleObject is called, which could cause a memory leak if the child classes
(Door, etc.) have something to release in their destructor.